+++ src/base/ContainerInterface.php
@@ -20,9 +20,7 @@ interface ContainerInterface
 {
     // Constants
     // =========================================================================
-
-    public const CONTAINER_TYPE = 'GenericContainer';
-
+    
     /**
      * @event IncludeContainerEvent The event that is triggered when a container is
      * about to be included.