+++ src/controllers/SettingsController.php
@@ -12,6 +12,7 @@ namespace nystudio107\seomatic\controllers;
 use Craft;
 use craft\elements\Asset;
 use craft\errors\MissingComponentException;
+use craft\helpers\Cp;
 use craft\helpers\UrlHelper;
 use craft\web\Controller;
 use craft\web\UrlManager;
@@ -107,6 +108,7 @@ class SettingsController extends Controller
     {
         $variables = [];
         // Get the site to edit
+        $siteHandle = $this->getCpSiteHandle($siteHandle);
         $siteId = $this->getSiteIdFromHandle($siteHandle);
         $pluginName = Seomatic::$settings->pluginName;
         $templateTitle = Craft::t('seomatic', 'Dashboard');
@@ -230,6 +232,7 @@ class SettingsController extends Controller
     public function actionGlobal(string $subSection = 'general', string $siteHandle = null, $loadFromSiteHandle = null, $editedMetaBundle = null): Response
     {
         $variables = [];
+        $siteHandle = $this->getCpSiteHandle($siteHandle);
         $siteId = $this->getSiteIdFromHandle($siteHandle);
 
         $pluginName = Seomatic::$settings->pluginName;
@@ -448,6 +451,7 @@ class SettingsController extends Controller
     {
         $variables = [];
         // Get the site to edit
+        $siteHandle = $this->getCpSiteHandle($siteHandle);
         $siteId = $this->getSiteIdFromHandle($siteHandle);
 
         $pluginName = Seomatic::$settings->pluginName;
@@ -514,6 +518,7 @@ class SettingsController extends Controller
         $variables = [];
         // @TODO: Let people choose an entry/categorygroup/product as the preview
         // Get the site to edit
+        $siteHandle = $this->getCpSiteHandle($siteHandle);
         $siteId = $this->getSiteIdFromHandle($siteHandle);
         if (is_string($typeId)) {
             $typeId = (int)$typeId;
@@ -526,11 +531,17 @@ class SettingsController extends Controller
         }
         $variables['typeMenu'] = $typeMenu;
         $variables['currentTypeId'] = null;
-        if (!empty($typeMenu)) {
+        $variables['specificTypeId'] = null;
+        if (count($typeMenu) > 1) {
             $currentType = reset($typeMenu);
             $variables['currentType'] = $typeMenu[$typeId] ?? $currentType;
             $variables['currentTypeId'] = $typeId ?? key($typeMenu);
             $typeId = (int)$variables['currentTypeId'];
+        }
+        // If there's only one EntryType, don't bother displaying the menu
+        if (count($typeMenu) === 1) {
+            $variables['typeMenu'] = [];
+            $variables['specificTypeId'] = $typeId ?? key($typeMenu);
         }
         $pluginName = Seomatic::$settings->pluginName;
         // Asset bundle
@@ -641,6 +652,7 @@ class SettingsController extends Controller
         $sourceHandle = $request->getParam('sourceHandle');
         $siteId = $request->getParam('siteId');
         $typeId = $request->getParam('typeId') ?? null;
+        $specificTypeId = $request->getParam('specificTypeId') ?? null;
         $globalsSettings = $request->getParam('metaGlobalVars');
         $bundleSettings = $request->getParam('metaBundleSettings');
         $sitemapSettings = $request->getParam('metaSitemapVars');
@@ -679,6 +691,12 @@ class SettingsController extends Controller
             Seomatic::$plugin->metaBundles->syncBundleWithConfig($metaBundle, true);
             $metaBundle->typeId = $typeId;
             Seomatic::$plugin->metaBundles->updateMetaBundle($metaBundle, $siteId);
+            // If there's also a specific typeId associated with this section, save the same
+            // metabundle there too, to fix: https://github.com/nystudio107/craft-seomatic/issues/1557
+            if (!empty($specificTypeId)) {
+                $metaBundle->typeId = $specificTypeId;
+                Seomatic::$plugin->metaBundles->updateMetaBundle($metaBundle, $siteId);
+            }
 
             Seomatic::$plugin->clearAllCaches();
             Craft::$app->getSession()->setNotice(Craft::t('seomatic', 'SEOmatic content settings saved.'));
@@ -702,6 +720,7 @@ class SettingsController extends Controller
     {
         $variables = [];
         // Get the site to edit
+        $siteHandle = $this->getCpSiteHandle($siteHandle);
         $siteId = $this->getSiteIdFromHandle($siteHandle);
 
         $pluginName = Seomatic::$settings->pluginName;
@@ -901,6 +920,7 @@ class SettingsController extends Controller
     {
         $variables = [];
         // Get the site to edit
+        $siteHandle = $this->getCpSiteHandle($siteHandle);
         $siteId = $this->getSiteIdFromHandle($siteHandle);
         // Enabled sites
         $this->setMultiSiteVariables($siteHandle, $siteId, $variables);
@@ -1086,6 +1106,22 @@ class SettingsController extends Controller
 
     // Protected Methods
     // =========================================================================
+
+    /**
+     * @param $siteHandle
+     * @return string|null
+     */
+    protected function getCpSiteHandle($siteHandle)
+    {
+        // As of Craft 4, the site query parameter is appended to CP urls to indicate the current
+        // site that is being edited, so respect it
+        $cpSite = Cp::requestedSite();
+        if ($cpSite) {
+            return $cpSite->handle;
+        }
+
+        return $siteHandle;
+    }
 
     /**
      * Return a siteId from a siteHandle